fix(diff-overlay): stop the header clipping its own controls
Everything on the diff header except the commit subject was
`flex_shrink_0`, including a focused file's path -- which is unbounded.
In a docked column the subject was squeezed to nothing and the fixed
elements then overflowed, clipping away the view switch and the close
tile: the only two things on the strip you can actually click.

The path and the byline yield first now (`flex_shrink(999.)` against the
subject's `1.`), because both have a second home -- the path one line
down in the file list, the author in the commit list -- and the subject
has none. The path shrinks head-first via `split_path_leaf`, so the
filename is the last thing to go rather than the first.

`split_path_leaf` moves from `right_panel` to `path_display`, with its
tests, since that module is where a path's display spelling lives.
